+++ Front-End/src/app/listados/departamentos/departamentos.component.ts
@@ -4,7 +4,7 @@ import { DepartamentoService } from '../../services/departamento.service';
 @Component({
   selector: 'app-departamentos',
   templateUrl: './departamentos.component.html',
-  styleUrl: './departamentos.component.css'
+  styleUrls: ['./departamentos.component.css']
 })
 export class DepartamentosComponent implements OnInit {
   departamentos: any[] = [];
@@ -18,8 +18,11 @@ export class DepartamentosComponent implements OnInit {
   }
 
   eliminarDepartamento(id: string): void {
+    const confirmacion = window.confirm('¿Estás seguro de que deseas eliminar este departamento?');
+    if (confirmacion) {
       this.departamentoService.eliminarDepartamento(id).subscribe(() => {
           this.departamentos = this.departamentos.filter(emp => emp._id !== id);
       });
+    }
   }
 }
