+++ Front-End/src/app/listados/parentescos/parentescos.component.ts
@@ -4,7 +4,7 @@ import { ParentescoService } from '../../services/parentesco.service';
 @Component({
   selector: 'app-parentescos',
   templateUrl: './parentescos.component.html',
-  styleUrl: './parentescos.component.css'
+  styleUrls: ['./parentescos.component.css']
 })
 export class ParentescosComponent implements OnInit {
   parentescos: any[] = [];
@@ -18,8 +18,11 @@ export class ParentescosComponent implements OnInit {
   }
 
   eliminarParentesco(id: string): void {
+    const confirmacion = window.confirm('¿Estás seguro de que deseas eliminar este parentesco?');
+    if (confirmacion) {
       this.parentescoService.eliminarParentesco(id).subscribe(() => {
           this.parentescos = this.parentescos.filter(emp => emp._id !== id);
       });
+    }
   }
 }
